+++ src/components/auth/forgot-password_container.js
@@ -5,8 +5,8 @@ import { withTranslation } from 'react-i18next';
 import Link from '@/components/link';
 import Button from '@/components/button';
 import Loading from '@/components/loading';
-import Validators from '@/components/form/validators';
-import Form from '@/components/new-form/form';
+import Validators from '@/components/old-form/validators';
+import Form from '@/components/form';
 import Input from '@/components/input';
 import { actions as notificationActions } from '@/components/notification';
 import { actions as storeActions } from '@/store';
