+++ src/components/company-settings/general/index.js
@@ -6,12 +6,12 @@ import Modal from '@/components/section-modal';
 import Button from '@/components/button';
 import Navigation from '@/components/navigation';
 import OrganizationIcon from '@/components/organization-icon';
-import Validators from '@/components/form/validators';
-import File from '@/components/form/file';
-import Input from '@/components/form/input';
+import Validators from '@/components/old-form/validators';
+import File from '@/components/old-form/file';
+import Input from '@/components/old-form/input';
 import { api } from '@';
 import { withRouter } from '@/hoc';
-import { actions as formActions } from '@/components/form';
+import { actions as formActions } from '@/components/old-form';
 import { actions as notificationActions } from '@/components/notification';
 import { actions as organizationsActions } from '@/store/organizations';
 import { withTranslation } from 'react-i18next';
