+++ src/components/old-form/index.js
@@ -1,0 +1,2 @@
+export {default as actions} from './actions';
+export {default as reducers} from './reducers';
