+++ src/components/old-form/input/index.js
@@ -3,8 +3,8 @@ import { connect } from 'react-redux';
 import get from 'lodash/get';
 import find from 'lodash/find';
 import InputField from '@/components/input';
-import { actions as formActions } from '@/components/form';
-import Validators from '@/components/form/validators';
+import { actions as formActions } from '@/components/old-form';
+import Validators from '@/components/old-form/validators';
 
 class Input extends Component {
   onInput = event => {
