+++ src/components/section_container/change-password/index.js
@@ -3,11 +3,11 @@ import compose from 'recompose/compose';
 import { connect } from 'react-redux';
 import Modal from '@/components/section-modal';
 import Loading from '@/components/loading';
-import Validators from '@/components/form/validators';
-import Form from '@/components/form/form';
-import Input from '@/components/form/input';
+import Validators from '@/components/old-form/validators';
+import Form from '@/components/old-form/form';
+import Input from '@/components/old-form/input';
 import { api } from '@';
-import { actions as formActions } from '@/components/form';
+import { actions as formActions } from '@/components/old-form';
 import { actions as notificationActions } from '@/components/notification';
 import { withTranslation } from 'react-i18next';
 import style from './style.css';
