+++ src/components/section_container/edit-profile/index.js
@@ -5,14 +5,14 @@ import { connect } from 'react-redux';
 import Modal from '@/components/section-modal';
 import Avatar from '@/components/avatar';
 import Loading from '@/components/loading';
-import Validators from '@/components/form/validators';
-import Form from '@/components/form/form';
-import Input from '@/components/form/input';
-import Checkbox from '@/components/form/checkbox';
-import File from '@/components/form/file';
+import Validators from '@/components/old-form/validators';
+import Form from '@/components/old-form/form';
+import Input from '@/components/old-form/input';
+import Checkbox from '@/components/old-form/checkbox';
+import File from '@/components/old-form/file';
 import { api } from '@';
 import { copy } from '@/helpers';
-import { actions as formActions } from '@/components/form';
+import { actions as formActions } from '@/components/old-form';
 import { actions as notificationActions } from '@/components/notification';
 import modalActions from '@/components/modal_container/actions';
 import { withTranslation } from 'react-i18next';
